move DTIBasedMeasure to new clinica.utils.dwi module
